Create a function that prints and returns the last digit of an integer

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -7,7 +7,7 @@
 
 int print_last_digit(int a)
 {
-	int i;
+	long i;
 
 	i = a;
 	if (i < 0)
